Examples

PinCFlow.Examples — Module

Examples

Module that defines PinCFlow.jl's example functions.

See also

PinCFlow.Examples.WavePacketTools

PinCFlow.Macros

PinCFlow.Types

PinCFlow.Integration

PinCFlow.PinCFlow

PinCFlow.Examples.cold_bubble — Function

cold_bubble(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "cold_bubble.h5",
    plot_file::AbstractString = "cold_bubble.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the cold-bubble example simulation.

Keywords

display_figure: Switch for displaying the figure created by plot_output.

npx: Number of MPI subdomains in $\hat{x}$.

npy: Number of MPI subdomains in $\hat{y}$.

npz: Number of MPI subdomains in $\hat{z}$.

output_file: File to which to save the output data.

plot_file: File to which to save the figure created by plot_output.

prepare_restart: Switch for writing all fields necessary for a restart.

visualize: Switch for visualizing the output with plot_output.

x_size: Number of grid points in $\hat{x}$.

y_size: Number of grid points in $\hat{y}$.

z_size: Number of grid points in $\hat{z}$.

PinCFlow.Examples.hot_bubble — Function

hot_bubble(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "hot_bubble.h5",
    plot_file::AbstractString = "hot_bubble.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the hot-bubble example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.mountain_wave — Function

mountain_wave(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "mountain_wave.h5",
    plot_file::AbstractString = "mountain_wave.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the mountain-wave example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.periodic_hill — Function

periodic_hill(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "periodic_hill.h5",
    plot_file::AbstractString = "periodic_hill.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the periodic-hill example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.vortex — Function

vortex(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "vortex.h5",
    plot_file::AbstractString = "vortex.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the vortex example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wave_packet — Function

wave_packet(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wave_packet.h5",
    plot_file::AbstractString = "wave_packet.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the wave-packet example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wkb_mountain_wave — Function

wkb_mountain_wave(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wkb_mountain_wave.h5",
    plot_file::AbstractString = "wkb_mountain_wave.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the WKB-mountain-wave example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wkb_wave_packet — Function

wkb_wave_packet(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wkb_wave_packet.h5",
    plot_file::AbstractString = "wkb_wave_packet.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the WKB-wave-packet example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

WavePacketTools

PinCFlow.Examples.WavePacketTools — Module

WavePacketTools

Module that defines helper functions for setting up the wave-packet and WKB-wave-packet example simulations.

See also

PinCFlow.Macros

PinCFlow.Types

PinCFlow.Examples.WavePacketTools.bhat — Function

bhat(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the buoyancy wave amplitude at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.envelope

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.envelope — Function

envelope(parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the circular cosine envelope at $\left(x, y, z\right)$.

Arguments

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

PinCFlow.Examples.WavePacketTools.ijk — Function

ijk(state::State, x::Real, y::Real, z::Real)::CartesianIndex

Return the indices of the grid point in state that is nearest to $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

x: Zonal position.

y: Meridional position.

z: Vertical position.

PinCFlow.Examples.WavePacketTools.n2 — Function

n2(state::State, x::Real, y::Real, z::Real)::Real

Return the squared buoyancy frequency of state at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.ijk

PinCFlow.Examples.WavePacketTools.omega — Function

omega(
    state::State,
    parameters::NamedTuple,
    x::Real,
    y::Real,
    z::Real,
)::Real

Return the intrinsic frequency at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state (needed by n2).

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.phi — Function

envelope(parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the phase at $\left(x, y, z\right)$.

Arguments

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

PinCFlow.Examples.WavePacketTools.pihat — Function

pihat(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the Exner-pressure wave amplitude at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.bhat

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.omega

PinCFlow.Examples.WavePacketTools.thetabar

PinCFlow.Examples.WavePacketTools.rhobar — Function

rhobar(state::State, x::Real, y::Real, z::Real)::Real

Return the reference density of state at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.ijk

PinCFlow.Examples.WavePacketTools.thetabar — Function

thetabar(state::State, x::Real, y::Real, z::Real)::Real

Return the reference potential temperature of state at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.ijk

PinCFlow.Examples.WavePacketTools.uhat — Function

uhat(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the zonal-wind wave amplitude at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.bhat

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.omega

PinCFlow.Examples.WavePacketTools.vhat — Function

vhat(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the meridional-wind wave amplitude at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.bhat

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.omega

PinCFlow.Examples.WavePacketTools.wave_action_density — Function

wave_action_density(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the wave-action density at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.bhat

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.omega

PinCFlow.Examples.WavePacketTools.rhobar

PinCFlow.Examples.WavePacketTools.what — Function

what(state::State, parameters::NamedTuple, x::Real, y::Real, z::Real)::Real

Return the vertical-wind wave amplitude at $\left(x, y, z\right)$.

Arguments

state: Auxiliary model state.

parameters: Parameters of the wave-packet configuration.

x: Zonal position.

y: Meridional position.

z: Vertical position.

See also

PinCFlow.Examples.WavePacketTools.bhat

PinCFlow.Examples.WavePacketTools.n2

PinCFlow.Examples.WavePacketTools.omega