Public API

PinCFlow — Module

PinCFlow

Main module of PinCFlow.jl.

See also

PinCFlow.Types

PinCFlow.Integration

Namelists

PinCFlow.Types.NamelistTypes.AtmosphereNamelist — Type

AtmosphereNamelist

Namelist for parameters used in the definition of the atmospheric background and the initialization of prognostic variables.

AtmosphereNamelist(;
    model::Symbol = :PseudoIncompressible,
    specify_reynolds_number::Bool = false,
    inverse_reynolds_number::Real = 0.0E+0,
    kinematic_viscosity::Real = 1.5E-5,
    thermal_conductivity::Real = 3.0E-5,
    kinematic_diffusivity::Real = 0.0E+0,
    background::Symbol = :Isothermal,
    buoyancy_frequency::Real = 1.0E-2,
    potential_temperature::Real = 3.0E+2,
    temperature::Real = 3.0E+2,
    ground_pressure::Real = 1.0E+5,
    coriolis_frequency::Real = 1.0E-4,
    tropopause_height::Real = 1.0E+4,
    troposphere_lapse_rate::Real = 6.5E-3,
    stratosphere_lapse_rate::Real = -5.0E-3,
    initial_rhop::Function = (x, y, z) -> 0.0,
    initial_u::Function = (x, y, z) -> 0.0,
    initial_v::Function = (x, y, z) -> 0.0,
    initial_w::Function = (x, y, z) -> 0.0,
    initial_pip::Function = (x, y, z) -> 0.0,
    initial_thetap::Function = (x, y, z) -> 0.0,
    buoyancy_initialization::Symbol = :initial_rhop,
)::AtmosphereNamelist

Construct an AtmosphereNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

model::Symbol: Dynamic equations.

specify_reynolds_number::Bool: Flag to specify inverse Reynolds number instead of viscosity.

inverse_reynolds_number::Float64: Inverse Reynolds number.

kinematic_viscosity::Float64: Kinematic viscosity at the surface.

thermal_conductivity::Float64: Thermal conductivity at the surface.

kinematic_diffusivity::Float64: Kinematic diffusivity at the surface.

background::Symbol: Atmospheric background.

buoyancy_frequency::Float64: Buoyancy frequency if background === :StableStratification.

potential_temperature::Float64: Reference potential temperature.

temperature::Float64: Reference temperature.

ground_pressure::Float64: Reference pressure.

coriolis_frequency::Float64: Coriolis frequency of the $f$-plane.

tropopause_height::Float64: Height of the tropopause for background === :Realistic or background === :LapseRates.

troposphere_lapse_rate::Float64: Lapse rate in the troposphere for background === :LapseRates.

stratosphere_lapse_rate::Float64: Lapse rate in the stratosphere for background === :LapseRates.

initial_rhop::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the density fluctuations if buoyancy_initialization === :initial_rhop.

initial_u::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the zonal wind.

initial_v::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the meridional wind.

initial_w::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the vertical wind.

initial_pip::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the Exner-pressure fluctuations.

initial_thetap::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the density fluctuations if buoyancy_initialization === :initial_thetap.

buoyancy_initialization::Symbol: Switch for using initial_rhop (if set to :initial_rhop) or initial_thetap (if set to :initial_thetap) to initialize the buoyancy.

PinCFlow.Types.NamelistTypes.DiscretizationNamelist — Type

DiscretizationNamelist

Namelist for parameters describing the discretization.

DiscretizationNamelist(;
    cfl_number::Real = 5.0E-1,
    wkb_cfl_number::Real = 3.0E-1,
    dtmin::Real = 1.0E-6,
    dtmax::Real = 1.0E+3,
    adaptive_time_step::Bool = true,
    limiter_type::Symbol = :MCVariant,
)::DiscretizationNamelist

Construct a DiscretizationNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

cfl_number::Float64: Number used for the CFL condition in the time step computation.

wkb_cfl_number::Float64: Number used for the WKB-CFL condition in the time step computation.

dtmin::Float64: Minimum time step allowed for the integration.

dtmax::Float64: Maximum time step allowed for the integration.

adaptive_time_step::Bool: Switch for using stability criteria to determine the time step. If set to false, dtmax is used as a fixed time step.

limiter_type::Symbol: Flux limiter used by the MUSCL scheme.

PinCFlow.Types.NamelistTypes.DomainNamelist — Type

DomainNamelist

Namelist for parameters describing the model domain.

DomainNamelist(;
    x_size::Integer = 1,
    y_size::Integer = 1,
    z_size::Integer = 1,
    nbx::Integer = 3,
    nby::Integer = 3,
    nbz::Integer = 3,
    lx::Real = 1.0E+3,
    ly::Real = 1.0E+3,
    lz::Real = 1.0E+3,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
)::DomainNamelist

Construct a DomainNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

x_size::Int: Number of grid cells in $\hat{x}$-direction.

y_size::Int: Number of grid cells in $\hat{y}$-direction.

z_size::Int: Number of grid cells in $\hat{z}$-direction.

nbx::Int: Number of boundary/halo cells in $\hat{x}$-direction.

nby::Int: Number of boundary/halo cells in $\hat{y}$-direction.

nbz::Int: Number of boundary/halo cells in $\hat{z}$-direction.

lx::Float64: Domain extent in $\hat{x}$-direction.

ly::Float64: Domain extent in $\hat{y}$-direction.

lz::Float64: Domain extent in $\hat{z}$-direction.

npx::Int: Number of MPI processes in $\hat{x}$-direction.

npy::Int: Number of MPI processes in $\hat{y}$-direction.

npz::Int: Number of MPI processes in $\hat{z}$-direction.

PinCFlow.Types.NamelistTypes.GridNamelist — Type

GridNamelist

Namelist for parameters describing the grid.

GridNamelist(;
    vertical_grid_stretching::Function = zhat -> zhat,
    resolved_topography::Function = (x, y) -> 0.0,
    unresolved_topography::Function = (alpha, x, y) -> (0.0, 0.0, 0.0),
)::GridNamelist

Construct a GridNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

vertical_grid_stretching::FunctionWrapper{Float64, Tuple{Float64}}: Function that defines the vertical grid stretching $\tilde{z} \left(\hat{z}\right)$.

resolved_topography::FunctionWrapper{Float64, NTuple{2, Float64}}: Function that returns the resolved topography at a specified horizontal position.

unresolved_topography::FunctionWrapper{NTuple{3, Float64}, Tuple{Int, Float64, Float64}}: Function that returns a specified spectral mode of the unresolved topography at a specified horizontal position.

PinCFlow.Types.NamelistTypes.OutputNamelist — Type

OutputNamelist

Namelist for I/O parameters.

OutputNamelist(;
    output_variables::Vector{Symbol} = Symbol[],
    save_ray_volumes::Bool = false,
    prepare_restart::Bool = false,
    restart::Bool = false,
    iin::Integer = -1,
    output_steps::Bool = false,
    nout::Integer = 1,
    iterations::Integer = 1,
    output_interval::Real = 3.6E+3,
    tmax::Real = 3.6E+3,
    input_file::AbstractString = "./pincflow_input.h5",
    output_file::AbstractString = "./pincflow_output.h5",
)::OutputNamelist

Construct an OutputNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

output_variables::Vector{Symbol}: A vector of symbols representing the variables that should be written to the output file.

save_ray_volumes::Bool: A boolean indicating whether to write ray-volume data.

prepare_restart::Bool: A boolean indicating whether to write all variables needed for restart simulations.

restart::Bool: A boolean indicating whether to initialize with data from a previous state (as written in input_file).

iin::Int: Temporal index in input_file at which to read the data to initialize with in restart simulations. If it's set to the default value -1, the data will be read at the last index.

output_steps::Bool: If set to true, write output every nout time steps.

nout::Int: Output interval (in indices) if output_steps == true.

iterations::Int: Maximum number of iterations if output_steps == true.

output_interval::Float64: Output interval (in physical time) if output_steps == false.

tmax::Float64: Simulation time if output_steps == false.

input_file::String: File from which to read input data in restart simulations.

output_file::String: File to which output data is written.

PinCFlow.Types.NamelistTypes.PoissonNamelist — Type

PoissonNamelist

Namelist for parameters used by the Poisson solver.

PoissonNamelist(;
    tolerance::Real = 1.0E-8,
    poisson_iterations::Integer = 1000,
    preconditioner::Bool = true,
    dtau::Real = 1.0E+0,
    preconditioner_iterations::Integer = 2,
    initial_cleaning::Bool = true,
    tolerance_is_relative::Bool = false,
)::PoissonNamelist

Construct a PoissonNamelists instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

tolerance::Float64: Tolerance for the convergence criterion of the Poisson solver.

poisson_iterations::Int: Maximum number of iterations performed by the Poisson solver before it terminates regardless of convergence.

preconditioner::Bool: Whether to use a preconditioner to accelerate the convergence of the Poisson solver.

dtau::Float64: Pseudo-time step coefficient used by the preconditioner.

preconditioner_iterations::Int: Number of iterations performed by the preconditioner.

initial_cleaning::Bool: Whether to solve the Poisson problem at initialization to guarantee an initially divergence-free state.

tolerance_is_relative::Bool: If set to true, the tolerance used for the convergence criterion is given by tolerance. If set to false, the tolerance is given by tolerance divided by a reference value determined from the right-hand side.

PinCFlow.Types.NamelistTypes.SpongeNamelist — Type

SpongeNamelist

Namelist for parameters describing the sponge.

SpongeNamelist(;
    damp_horizontal_wind_on_rhs::Bool = false,
    relax_to_mean::Bool = false,
    lhs_sponge::Function = (x, y, z, t, dt) -> 0.0,
    rhs_sponge::Function = (x, y, z, t, dt) -> 0.0,
    relaxed_u::Function = (x, y, z, t, dt) -> 0.0,
    relaxed_v::Function = (x, y, z, t, dt) -> 0.0,
    relaxed_w::Function = (x, y, z, t, dt) -> 0.0,
)::SpongeNamelist

Construct a SpongeNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

damp_horizontal_wind_on_rhs::Bool: Switch for applying the RHS sponge to the horizontal wind.

relax_to_mean::Bool: Switch for relaxing the wind towards its averages on the terrain-following surfaces. If set to false, the relaxation wind is computed with relaxed_u, relaxed_v and relaxed_w.

lhs_sponge::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the Rayleigh-damping coefficient of the LHS sponge.

rhs_sponge::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the Rayleigh-damping coefficient of the RHS sponge.

relaxed_u::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the zonal relaxation wind if relax_to_mean is set to false.

relaxed_v::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the meridional relaxation wind if relax_to_mean is set to false.

relaxed_w::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the vertical relaxation wind if relax_to_mean is set to false.

PinCFlow.Types.NamelistTypes.TracerNamelist — Type

TracerNamelist

Namelist for the inclusion of a tracer and the calculation of the leading-order gravity-wave impact.

TracerNamelist(;
    tracer_setup::Symbol = :NoTracer,
    leading_order_impact::Bool = false,
    initial_chi::Function = (x, y, z) -> 0.0,
    relaxed_chi::Function = (x, y, z, t, dt) -> 0.0,
    apply_lhs_sponge_to_tracer::Bool = true,
)::TracerNamelist

Construct a TracerNamelist instance with the given keyword arguments as properties.

Fields/Keywords

tracer_setup::Symbol: General tracer configuration.

leading_order_impact::Bool: Flag to include the leading-order impact of gravity waves when parameterizing waves with the WKB model.

initial_chi::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the tracer.

relaxed_chi::FunctionWrapper{Float64, NTuple{5, Float64}}: Function used to compute the background tracer.

apply_lhs_sponge_to_tracer::Bool: Flag to relax the tracer fields to relaxed_chi.

PinCFlow.Types.NamelistTypes.TurbulenceNamelist — Type

TurbulenceNamelist

Namelist for the inclusion of a turbulence parameterization and the turbulent diffusion of momentum, mass-weighted potential temperature, and tracers.

TurbulenceNamelist(;
    turbulence_scheme::Symbol = :TKEScheme,
    momentum_coupling::Bool = true,
    entropy_coupling::Bool = true,
    tracer_coupling::Bool = true,
    initial_tke::Function = (x, y, z) -> 5e-5,
)::TurbulenceNamelist

Construct a TurbulenceNamelist instance with the given keyword arguments as properties.

Fields/Keywords

turbulence_scheme::Symbol: General turbulence parameterization configuration.

momentum_coupling::Bool: Switch for turbulent diffusion of momentum.

entropy_coupling::Bool: Switch for turbulent diffusion of the mass-specific potential temperature.

tracer_coupling::Bool: Switch for turbulent diffusion of tracers.

initial_tke::FunctionWrapper{Float64, NTuple{3, Float64}}: Function used to initialize the mass-specific turbulent kinetic energy.

PinCFlow.Types.NamelistTypes.WKBNamelist — Type

WKBNamelist

Namelist for parameters used by MS-GWaM.

WKBNamelist(;
    nrx::Integer = 1,
    nry::Integer = 1,
    nrz::Integer = 1,
    nrk::Integer = 1,
    nrl::Integer = 1,
    nrm::Integer = 1,
    k_bins::Integer = 1,
    l_bins::Integer = 1,
    m_bins::Integer = 3,
    dkr_factor::Real = 1.0E-1,
    dlr_factor::Real = 1.0E-1,
    dmr_factor::Real = 1.0E-1,
    branch::Integer = -1,
    merge_mode::Symbol = :ConstantWaveAction,
    filter_order::Integer = 2,
    smooth_tendencies::Bool = true,
    filter_type::Symbol = :ShapiroFilter,
    impact_altitude::Real = 0.0E+0,
    use_saturation::Bool = true,
    saturation_threshold::Real = 1.0E+0,
    wkb_mode::Symbol = :NoWKB,
    orographic_flow::Symbol = :Average,
    blocking::Bool = false,
    long_threshold::Real = 2.5E-1,
    drag_coefficient::Real = 1.0E+0,
    reduction_coefficient::Real = 1.0E+0,
    wave_modes::Integer = 1,
    initial_wave_field::Function = (alpha, x, y, z) ->
        (0.0, 0.0, 0.0, 0.0, 0.0),
    elastic_mode_selection::Bool = false,
    minimum_mode_count::Integer = wave_modes,
    maximum_mode_count::Integer = wave_modes,
    minimum_power_fraction::Real = 1.0E+0,
    maximum_power_fraction::Real = 1.0E+0,
    turbulent_damping::Bool = false,
)::WKBNamelist

Construct a WKBNamelist instance with the given keyword arguments as properties, converting them to meet the type constraints.

Fields/Keywords

nrx::Int: Number of ray-volumes launched per grid cell and wave mode in $\hat{x}$-direction.

nry::Int: Number of ray-volumes launched per grid cell and wave mode in $\hat{y}$-direction.

nrz::Int: Number of ray-volumes launched per grid cell and wave mode in $\hat{z}$-direction.

nrk::Int: Number of ray-volumes launched per grid cell and wave mode in $k$-direction.

nrl::Int: Number of ray-volumes launched per grid cell and wave mode in $l$-direction.

nrm::Int: Number of ray-volumes launched per grid cell and wave mode in $m$-direction.

k_bins::Int: Number of $k$-bins in the merging algorithm for single-column/multi-column mode (must be an odd number).

l_bins::Int: Number of $l$-bins in the merging algorithm for single-column/multi-column mode (must be an odd number).

m_bins::Int: Number of $m$-bins in the merging algorithm for single-column/multi-column mode (must be an odd number).

dkr_factor::Float64: Relative initial ray-volume extent in $k$.

dlr_factor::Float64: Relative initial ray-volume extent in $l$.

dmr_factor::Float64: Relative initial ray-volume extent in $m$.

branch::Int: Frequency branch.

merge_mode::Symbol: Ray-volume merging strategy (conserved quantity).

filter_order::Int: Order of the smoothing applied to the mean-flow tendencies.

smooth_tendencies::Bool: Switch for smoothing the mean-flow tendencies.

filter_type::Symbol: Filter to use for the smoothing of the mean-flow tendencies.

impact_altitude::Float64: Minimum altitude for ray-tracing and mean-flow impact.

use_saturation::Bool: Switch for the saturation scheme.

saturation_threshold::Float64: Relative saturation threshold.

wkb_mode::Symbol: Approximations used by MS-GWaM.

orographic_flow::Symbol: Method used to approximate quantities of the mountain-wave-generating background flow.

blocking::Bool: Switch for parameterizing blocking in WKB-mountain-wave simulations.

long_threshold::Float64: Long-number threshold used by the blocked-layer scheme.

drag_coefficient::Float64: Dimensionless drag coefficient used by the blocked-layer scheme.

reduction_coefficient::Float64: Dimensionless coefficient that regulates how much of the blocked layer contributes to a reduction of the wave amplitude.

wave_modes::Int: Number of wave modes per grid cell.

initial_wave_field::FunctionWrapper{NTuple{5, Float64}, Tuple{Int, Float64, Float64, Float64}}: Function used to set the initial wavenumbers, intrinsic frequency and wave-action density of each wave mode.

elastic_mode_selection::Bool: Switch for elastic mode selection in ray-volume sources.

minimum_mode_count::Int: Minimum number of modes selected by the elastic-mode-selection algorithm.

minimum_mode_count::Int: Maximum number of modes selected by the elastic-mode-selection algorithm.

minimum_power_fraction::Float64: Minimum power fraction retained by the elastic-mode-selection algorithm.

maximum_power_fraction::Float64: Maximum power fraction retained by the elastic-mode-selection algorithm.

turbulent_damping::Bool: Damping of wave-action density due to turbulence.

Experimental

The blocked-layer scheme is an experimental feature that hasn't been fully validated yet.

Experimental

The turbulent damping of wave-action density is an experimental feature that hasn't been validated yet.

Experimental

The elastic mode selection is an experimental feature adapted from Banerjee (2026).

PinCFlow.Types.NamelistTypes.Namelists — Type

Namelists{
    A <: DomainNamelist,
    B <: OutputNamelist,
    C <: DiscretizationNamelist,
    D <: PoissonNamelist,
    E <: AtmosphereNamelist,
    F <: GridNamelist,
    G <: SpongeNamelist,
    H <: WKBNamelist,
    I <: TracerNamelist,
    K <: TurbulenceNamelist,
}

Collection of all configurable model parameters.

Namelists(;
    domain::DomainNamelist = DomainNamelist(),
    output::OutputNamelist = OutputNamelist(),
    discretization::DiscretizationNamelist = DiscretizationNamelist(),
    poisson::PoissonNamelist = PoissonNamelist(),
    atmosphere::AtmosphereNamelist = AtmosphereNamelist(),
    grid::GridNamelist = GridNamelist(),
    sponge::SpongeNamelist = SpongeNamelist(),
    wkb::WKBNamelist = WKBNamelist(),
    tracer::TracerNamelist = TracerNamelist(),
    turbulence::TurbulenceNamelist = TurbulenceNamelist(),
)::Namelists

Construct a Namelists instance with the given keyword arguments as properties.

Fields/Keywords

domain::A: Namelist for parameters describing the model domain.

output::B: Namelist for I/O parameters.

discretization::C: Namelist for parameters describing discretization.

poisson::D: Namelist for parameters used by the Poisson solver.

atmosphere::E: Namelist for parameters describing the atmospheric background.

grid::F: Namelist for parameters describing the grid.

sponge::G: Namelist for parameters describing the sponge.

wkb::H: Namelist for parameters used by MS-GWaM.

tracer::I: Namelist for parameters configuring tracer transport.

turbulence::K: Namelist for parameters configuring turbulence parameterization.

See also

PinCFlow.Types.NamelistTypes.DomainNamelist

PinCFlow.Types.NamelistTypes.OutputNamelist

PinCFlow.Types.NamelistTypes.DiscretizationNamelist

PinCFlow.Types.NamelistTypes.PoissonNamelist

PinCFlow.Types.NamelistTypes.AtmosphereNamelist

PinCFlow.Types.NamelistTypes.GridNamelist

PinCFlow.Types.NamelistTypes.SpongeNamelist

PinCFlow.Types.NamelistTypes.WKBNamelist

PinCFlow.Types.NamelistTypes.TracerNamelist

PinCFlow.Types.NamelistTypes.TurbulenceNamelist

Types

PinCFlow.Types.State — Type

State{
    A <: Namelists,
    B <: Time,
    C <: Constants,
    D <: Domain,
    E <: Grid,
    F <: Atmosphere,
    G <: Sponge,
    H <: Poisson,
    I <: Variables,
    J <: WKB,
    K <: Tracer,
    L <: Turbulence,
}

Model state container.

An instance of this composite type holds complete information about the model configuration and simulation state, so that it is sufficient as primary input to most methods. The construction of such an instance is the first operation performed in PinCFlow.Integration.integrate, since it almost fully initializes the model.

State(namelists::Namelists; base_comm::MPI.Comm = MPI.COMM_WORLD)::State

Construct a State instance and thus initialize the model.

This method first uses the parameters specified in namelists to construct instances of the composite types defined in FoundationalTypes (i.e. Constants, Time, Domain, Grid, Atmosphere and Sponge). It then uses these instances to prepare the arrays needed for the Poisson solver, the time integration and the parameterization of unresolved gravity waves with MS-GWaM. Afterwards, only three operations of the initialization process remain (these are performed by PinCFlow.Integration.integrate), namely the initial cleaning, the setting of the initial ray-volume properties and the reading of input data in restart simulations.

Fields

namelists::A: Namelists with all model parameters.

time::B: Runge-Kutta time integration coefficients.

constants::C: Physical constants and reference values.

domain::D: Collection of domain-decomposition and MPI-communication parameters.

grid::E: Collection of parameters and fields that describe the grid.

atmosphere::F: Atmospheric-background fields.

sponge::G: Sponge parameters and damping coefficients.

poisson::H: Workspace and solution arrays for the Poisson solver.

variables::I: Arrays needed for the predictions of the prognostic variables.

wkb::J: Container for WKB ray-tracing data and parameters.

tracer::K: Tracer setup and parameters.

turbulence::L: Turbulence setup, parameters, and turbulence energies.

Arguments

namelists: Namelists with all model parameters.

Keywords

base_comm: MPI communicator which is used to create the Cartesian communicator for the integration.

See also

PinCFlow.Types.FoundationalTypes.Constants

PinCFlow.Types.FoundationalTypes.Time

PinCFlow.Types.FoundationalTypes.Domain

PinCFlow.Types.FoundationalTypes.Grid

PinCFlow.Types.FoundationalTypes.Atmosphere

PinCFlow.Types.FoundationalTypes.Sponge

PinCFlow.Types.PoissonTypes.Poisson

PinCFlow.Types.VariableTypes.Variables

PinCFlow.Types.WKBTypes.WKB

PinCFlow.Types.TracerTypes.Tracer

PinCFlow.Types.TurbulenceTypes.Turbulence

Integration

PinCFlow.Integration.integrate — Function

integrate(
    namelists::Vararg{Namelists};
    base_comm::MPI.Comm = MPI.COMM_WORLD,
    delay::Real = 0,
)

Call the integrate method below and reduce potential exceptions with reduce_exceptions.

If length(namelists) == 1, the integrate method below is called with integrate(namelists[1], ParallelExceptions(); base_comm), otherwise, it is called with integrate(namelists[member], ParallelExceptions(); base_comm = member_comm), where member is an integer that distinguishes the members of an ensemble, and member_comm is the corresponding sub-communicator of base_comm.

integrate(
    namelists::Namelists,
    ::ParallelExceptions;
    base_comm::MPI.Comm = MPI.COMM_WORLD,
)

Initialize the model state and integrate it in time.

This method performs the complete time integration of the governing equations, using a semi-implicit time stepping scheme. It handles initialization, time stepping and output of the simulation data.

The initialization begins with the construction of the model state (an instance of the composite type State), which involves the setup of the MPI parallelization and the definition of all arrays that are needed repeatedly during the simulation. This is followed by an (optional) initial cleaning, in which the Poisson solver is called to ensure that the initial dynamic fields satisfy the divergence constraint imposed by the thermodynamic energy equation. Afterwards, the initialization of MS-GWaM is completed by adding ray volumes to the previously defined arrays. If the simulation is supposed to start from a previous model state, the fields are then overwritten with the data in the corresponding input file. Finally, the output file is created and the initial state is written into it.

At the beginning of each time-loop iteration, the time step is determined from several stability criteria, using compute_time_step. In case the updated simulation time is later than the next output time, the time step is corrected accordingly. Subsequently, the damping coefficients of the sponges (which may depend on the time step) are calculated. Following this, MS-GWaM updates the unresolved gravity-wave field and computes the corresponding mean-flow impact. Then, the diffusion of momentum, mass-weighted potential temperature and tracers is applied. Afterwards, the resolved flow is updated in a semi-implicit time step, comprised of the following stages.

Explicit RK3 integration of LHS over $\Delta t / 2$.

Implicit Euler integration of RHS over $\Delta t / 2$.

Explicit Euler integration of RHS over $\Delta t / 2$.

Explicit RK3 integration of LHS over $\Delta t$.

Implicit Euler integration of RHS over $\Delta t / 2$.

Therein, the left-hand sides of the equations include advective fluxes, diffusion terms, rotation and heating, whereas the pressure gradient, buoyancy term and momentum-flux divergence due to unresolved gravity waves are on the right-hand sides. Boundary conditions are enforced continuously. At the end of the time step, the updated fields are written into the output file if the next output time has been reached.

In the case of turbulence parameterization, the turbulence variables are integrated after step 2.

Arguments

namelists: Namelists with all model parameters.

Keywords

base_comm: MPI communicator which is used to create the Cartesian communicator for the integration.

delay: Delay (in seconds) between the first exception and the following MPI abort.

See also

PinCFlow.Types.State

PinCFlow.Integration.modify_compressible_wind!

PinCFlow.Boundaries.set_boundaries!

PinCFlow.PoissonSolver.apply_corrector!

PinCFlow.Output.create_output

PinCFlow.Output.write_output

PinCFlow.MSGWaM.RayUpdate.initialize_rays!

PinCFlow.Output.read_input!

PinCFlow.Integration.synchronize_compressible_atmosphere!

PinCFlow.Integration.compute_time_step

PinCFlow.Update.compute_sponges!

PinCFlow.Integration.wkb_integration!

PinCFlow.Integration.synchronize_density_fluctuations!

PinCFlow.Integration.explicit_integration!

PinCFlow.Integration.implicit_integration!

PinCFlow.Integration.reset_predictands!

PinCFlow.Update.turbulent_diffusion!

PinCFlow.Integration.turbulence_integration!

Examples

PinCFlow.Examples.cold_bubble — Function

cold_bubble(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "cold_bubble.h5",
    plot_file::AbstractString = "cold_bubble.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the cold-bubble example simulation.

Keywords

display_figure: Switch for displaying the figure created by plot_output.

npx: Number of MPI subdomains in $\hat{x}$.

npy: Number of MPI subdomains in $\hat{y}$.

npz: Number of MPI subdomains in $\hat{z}$.

output_file: File to which to save the output data.

plot_file: File to which to save the figure created by plot_output.

prepare_restart: Switch for writing all fields necessary for a restart.

visualize: Switch for visualizing the output with plot_output.

x_size: Number of grid points in $\hat{x}$.

y_size: Number of grid points in $\hat{y}$.

z_size: Number of grid points in $\hat{z}$.

PinCFlow.Examples.hot_bubble — Function

hot_bubble(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "hot_bubble.h5",
    plot_file::AbstractString = "hot_bubble.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the hot-bubble example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.mountain_wave — Function

mountain_wave(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "mountain_wave.h5",
    plot_file::AbstractString = "mountain_wave.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the mountain-wave example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.periodic_hill — Function

periodic_hill(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "periodic_hill.h5",
    plot_file::AbstractString = "periodic_hill.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the periodic-hill example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.vortex — Function

vortex(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "vortex.h5",
    plot_file::AbstractString = "vortex.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the vortex example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wave_packet — Function

wave_packet(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wave_packet.h5",
    plot_file::AbstractString = "wave_packet.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the wave-packet example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wkb_mountain_wave — Function

wkb_mountain_wave(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wkb_mountain_wave.h5",
    plot_file::AbstractString = "wkb_mountain_wave.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the WKB-mountain-wave example simulation.

Keywords

The keywords are analogous to those of cold_bubble.

PinCFlow.Examples.wkb_wave_packet — Function

wkb_wave_packet(;
    display_figure::Bool = true,
    npx::Integer = 1,
    npy::Integer = 1,
    npz::Integer = 1,
    output_file::AbstractString = "wkb_wave_packet.h5",
    plot_file::AbstractString = "wkb_wave_packet.svg",
    prepare_restart::Bool = false,
    visualize::Bool = true,
    x_size::Integer = 20,
    y_size::Integer = 20,
    z_size::Integer = 20,
)

Run the WKB-wave-packet example simulation.

Keywords

The keywords are analogous to those of cold_bubble.