Macros

PinCFlow.Macros — Module

Macros

Module that contains PinCFlow.jl's convenience macros.

PinCFlow.Macros.dispatch — Function

dispatch(values::Expr, input::Any)::Any

If input is an expression, find the first dynamic value dispatch in it and make all dynamic value dispatches with the same argument static.

The dynamic value dispatches are made static by inserting input into an if block which compares the found argument to the options in values and replacing it accordingly in all Val calls.

Arguments

values: Expression of a tuple of allowed values.

input: Input object.

See also

PinCFlow.Macros.find_argument

PinCFlow.Macros.replace_argument

PinCFlow.Macros.find_argument — Function

find_argument(input::Any; argument::Any = nothing)::Any

If input is an expression, return the first occurrence of an argument to a Val call which itself is an argument to a function call.

If the described pattern is not found, this function returns nothing.

Arguments

input: Input object.

Keywords

argument: Argument found in input, to be communicated between recursive calls of this function. Note that if find_argument is called with argument set to any value other than nothing, it will always return that value.

PinCFlow.Macros.ivy — Function

ivy(x::Any; root::Bool = true)::Any

If x is an expression, climb its AST and apply @inbounds to the bodies of hard-local-scope blocks (i.e., macro and function definitions, let blocks, comprehensions, and generators), apply @inbounds and @views to x itself if root == true, and return it.

Arguments

x: Input expression

root: Switch for applying @inbounds and @views to x itself.

PinCFlow.Macros.replace_argument — Function

replace_argument(input::Any, argument::Any, literal::Any)::Any

Return a deep copy of input with all occurrences of argument as argument of a Val call replaced by literal (if input is an expression).

Arguments

input: Input object.

argument: Argument in Val calls.

literal: Replacement for argument in Val calls.

PinCFlow.Macros.@dispatch — Macro

@dispatch(values::Expr, input::Expr)

Find the first dynamic value dispatch in input and make all dynamic value dispatches with the same argument static.

The scope in which this macro is to be applied must include the argument of the first dynamic value dispatch, and the available methods of the involved functions must be covered by the options given in values.

Arguments

values: Expression of a tuple of allowed values.

input: Input expression with dynamic value dispatches.

See also

PinCFlow.Macros.dispatch

PinCFlow.Macros.@ivy — Macro

@ivy(x::Expr)

Return the expression x with all slices turned into views and all bounds checks turned off.

Arguments

x: Input expression.

See also

PinCFlow.Macros.ivy