PinCFlow

PinCFlow.PinCFlow — Module

PinCFlow

Main module of PinCFlow.jl.

See also

PinCFlow.Types

PinCFlow.Integration

PinCFlow.plot_contours — Function

plot_contours(
    file::AbstractString,
    data::HDF5.File,
    variable::AbstractString,
    indices::Vararg{NTuple{4, <:Integer}};
    number::Integer = 10,
    colormap_name::Symbol = :seismic,
    label::AbstractString = "",
)

Create contour plots of the dataset variable in data, display it and save it to file.

Arguments

file: File to save the plots to.

data: PinCFlow.jl output data.

variable: Output variable to plot.

indices: Tuples of four indices. The first three indices of each tuple define the planes in which the contours are to be plotted, whereas the fourth is the temporal index.

Keywords

number: Number of contour levels.

colormap_name: Colormap of choice.

label: Colorbar label for the plots.

PinCFlow.set_visualization_theme! — Function

set_visualization_theme!()

Configure Makie to use a customized theme.

PinCFlow.symmetric_contours — Function

symmetric_contours(
    minimum::AbstractFloat,
    maximum::AbstractFloat;
    number::Integer = 10,
    colormap_name::Symbol = :seismic,
)::Tuple{<:LinRange{<:AbstractFloat, <:Integer}, <:Any}

Compute symmetric contours levels and return them and a correspondingly indexed colormap.

Arguments

minimum: Smallest value to be plotted.

maximum: Largest value to be plotted.

Keywords

number: Number of contour levels.

colormap_name: Name under which the chosen colormap is registered.

PinCFlow.@ivy — Macro

@ivy(x::Expr)

Return the expression x with @inbounds and @views in front of it.