PinCFlowMakieExt

PinCFlowMakieExt.PinCFlowMakieExt — Module

PinCFlowMakieExt

Makie.jl extension (using the CairoMakie backend).

See also

PinCFlow.Macros

PinCFlow.plot_output — Function

plot_output(
    plot_file::AbstractString,
    data_file::AbstractString,
    fields::Vararg{
        Union{
            Tuple{Symbol, <:Integer},
            Tuple{Symbol, <:Real, <:Real, <:Real, <:Integer},
        },
    };
    colormap_name::Symbol = :seismic,
    color_tick_format::AbstractString = "{:.1E}",
    number::Integer = 10,
    space_unit::Symbol = :km,
    time_unit::Symbol = :h,
    x_tick_format::AbstractString = "{:.1f}",
    y_tick_format::AbstractString = "{:.1f}",
    z_tick_format::AbstractString = "{:.1f}",
)

Create contour plots of the dataset variable in data, display it and save it to file.

Arguments

plot_file: File to save the plots to.

data_file: HDF5 file with PinCFlow.jl output data.

fields: Either tuples of a variable name and a temporal index (for 2D data), or tuples of a variable name, three fractions which define the relative positions of the $\hat{y}$-$\hat{z}$, $\hat{x}$-$\hat{z}$, and $\hat{x}$-$\hat{y}$ planes, and a temporal index (for 3D data).

Keywords

colormap_name: Colormap of choice.

color_tick_format: Format string for the ticks of the colorbars.

display_figure: Switch for showing the figure with Makie.jl's display function.

number: Number of contour levels.

space_unit: Unit used for the coordinates. Must be :km or :m.

time_unit: Unit used for the time. Must be :d, :h, :min or :s.

x_tick_format: Format string for the ticks of the $x$-axis.

y_tick_format: Format string for the ticks of the $y$-axis.

z_tick_format: Format string for the ticks of the $z$-axis.

See also

PinCFlow.set_visualization_theme!

PinCFlowMakieExt.add_scatter_plot!

PinCFlowMakieExt.add_contour_plot!

PinCFlow.set_visualization_theme! — Function

set_visualization_theme!()

Configure Makie.jl to use a customized theme.

PinCFlow.symmetric_contours — Function

symmetric_contours(
    minimum::AbstractFloat,
    maximum::AbstractFloat;
    number::Integer = 10,
    colormap_name::Symbol = :seismic,
)::Tuple{<:LinRange{<:AbstractFloat, <:Integer}, <:Any}

Compute symmetric contour levels and return them and a correspondingly indexed colormap.

Arguments

minimum: Smallest value to be plotted.

maximum: Largest value to be plotted.

Keywords

number: Number of contour levels.

colormap_name: Name under which the chosen colormap is registered.

PinCFlowMakieExt.add_contour_plot! — Function

add_contour_plot!(figure::Figure, input::NamedTuple)

Add a contour plot to figure using the data in input.

Arguments

figure: Figure to modify.

input: Data and specifications for the plot.

See also

PinCFlow.symmetric_contours

PinCFlowMakieExt.tick_indices

PinCFlowMakieExt.add_scatter_plot! — Function

add_scatter_plot!(figure::Figure, input::NamedTuple)

Add a scatter plot to figure using the data in input.

Arguments

figure: Figure to modify.

input: Data and specifications for the plot.

See also

PinCFlow.symmetric_contours

PinCFlowMakieExt.tick_indices

PinCFlowMakieExt.tick_indices — Function

tick_indices(levels::AbstractVector{<:Real})::AbstractVector{<:Integer}

Return three symmetric indices for levels.

Arguments

levels: Abstract input vector for which the symmetric indices are determined.